% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rscripts.R
\name{backup_scripts}
\alias{backup_scripts}
\title{Title}
\usage{
backup_scripts(dir, backup_dir)
}
\arguments{
\item{dir}{folder holding the scripts}

\item{backup_dir}{folder the backup is written to}
}
\description{
Title
}
