% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{cheatmap}
\alias{cheatmap}
\title{Title}
\usage{
cheatmap(matrix, draw = T, show_clusters = NULL, add_cluster_n = T, ...)
}
\arguments{
\item{matrix}{numeric matrix to draw}

\item{draw}{draw the heatmap, or return the object unrendered}

\item{show_clusters}{annotate the cluster each row belongs to}

\item{add_cluster_n}{add the number of rows per cluster to the annotation}

\item{...}{additional arguments for ComplexHeatmap::Heatmap()}
}
\description{
Title
}
