% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{.extract_UniProt_seqinfo}
\alias{.extract_UniProt_seqinfo}
\title{Extract sequence-specific UniProt information from downloaded data}
\usage{
.extract_UniProt_seqinfo(
  data_UniProt,
  info_cols,
  rename_cols = c("name"),
  remove.na = T
)
}
\arguments{
\item{data_UniProt}{data frame of proteins and UniProt info as returned by
get_UniProt_data()}

\item{info_cols}{name of column to be extracted}

\item{rename_cols}{vector of names to rename extracted columns to (works
sequentially)}

\item{remove.na}{remove proteins without information}
}
\description{
Extract sequence-specific UniProt information from downloaded data
}
