% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{.identify_UniProt_infotype}
\alias{.identify_UniProt_infotype}
\title{Title}
\usage{
.identify_UniProt_infotype(vector_UniProt)
}
\arguments{
\item{vector_UniProt}{character vector of UniProt feature strings}
}
\description{
Title
}
