% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{.plot_peptides_on_seq}
\alias{.plot_peptides_on_seq}
\title{Title}
\usage{
.plot_peptides_on_seq(
  data_peptides,
  p,
  yvalue = "estimate",
  color = "regulation",
  color.scale = NULL,
  linewidth = "peptide_group",
  alpha = "peptide_group",
  Start = "Start",
  End = "End",
  add.length = 0,
  title = "",
  base_size = 16,
  min_y_range = 1,
  protein_width = 0.6,
  protein_range,
  add.labels = F,
  label = "regulation",
  label.size = 5,
  nudge_x = 0,
  nudge_y = 0,
  hjust = 0,
  vjust = 0,
  direction = "both",
  min.segment.length = 2
)
}
\arguments{
\item{data_peptides}{}

\item{p}{}

\item{yvalue}{}

\item{color}{}

\item{color.scale}{}

\item{linewidth}{}

\item{alpha}{}

\item{Start}{}

\item{End}{}

\item{add.length}{}

\item{title}{}

\item{base_size}{}

\item{min_y_range}{}

\item{protein_width}{}

\item{protein_range}{}

\item{add.labels}{}

\item{label}{}

\item{label.size}{}

\item{nudge_x}{}

\item{nudge_y}{}

\item{hjust}{}

\item{vjust}{}

\item{direction}{}

\item{min.segment.length}{}
}
\description{
Title
}
