% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{.plot_protein_features}
\alias{.plot_protein_features}
\title{Title}
\usage{
.plot_protein_features(
  data_sites,
  p,
  name = "name",
  start = "from",
  end = "to",
  add.length = 0,
  merge_name_position = F,
  protein_width = 0.6,
  base_size = 16,
  multiple_scaling_factor = 8,
  legend.position = "bottom",
  legend_width = 4,
  colors_manual
)
}
\arguments{
\item{data_sites}{}

\item{p}{}

\item{name}{}

\item{start}{}

\item{end}{}

\item{add.length}{}

\item{merge_name_position}{}

\item{protein_width}{}

\item{base_size}{}

\item{multiple_scaling_factor}{}

\item{legend.position}{}

\item{legend_width}{}

\item{colors_manual}{}
}
\description{
Title
}
