% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{.plot_protein_seq}
\alias{.plot_protein_seq}
\title{Title}
\usage{
.plot_protein_seq(
  protein_range = c(1, 100),
  protein_width = 0.6,
  add.length = 0,
  x_lab = "protein",
  y_lab = "",
  color = "#b0b0b0",
  alpha = 0.5,
  base_size = 16
)
}
\arguments{
\item{protein_range}{residue range to show, as c(start, end)}

\item{protein_width}{height of the protein bar}

\item{add.length}{padding added to each end of the sequence axis}

\item{x_lab}{x axis label}

\item{y_lab}{y axis label}

\item{color}{column mapped to colour}

\item{alpha}{column mapped to transparency}

\item{base_size}{base font size of the theme}
}
\description{
Title
}
