% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{.rcol}
\alias{.rcol}
\title{Returns a randomly generated vector of colors sorted by hue}
\usage{
.rcol(n = 1, seed = 42)
}
\arguments{
\item{n}{}

\item{seed}{}
}
\description{
Returns a randomly generated vector of colors sorted by hue
}
