% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_interpro.R
\name{download_UniParc_sequence_features}
\alias{download_UniParc_sequence_features}
\title{Download sequence features data for UniParc sequences}
\usage{
download_UniParc_sequence_features(
  uniparc_id,
  keep.empty = F,
  separate.multiple.sites = T,
  split.position = T,
  return.sequence = F,
  silent = T
)
}
\arguments{
\item{uniparc_id}{UniParc Id (one at a time)}

\item{keep.empty}{Keep Ids without data?}

\item{separate.multiple.sites}{split disconnected features into multiple rows}

\item{split.position}{add columns 'from' and 'to'}

\item{return.sequence}{keep the protein sequence column in the output}

\item{silent}{Suppress messages?}
}
\description{
Download sequence features data for UniParc sequences
}
\examples{
\dontrun{
download_UniParc_sequence_features("UPI000004C26F") 
}
}
