% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_GO.R
\name{extract_UniProt_GO2TERM2GENE}
\alias{extract_UniProt_GO2TERM2GENE}
\title{Title}
\usage{
extract_UniProt_GO2TERM2GENE(
  data_UniProt_GO,
  entry_col = "Entry",
  GO_col = c("BP", "CC", "MF", "all", "ids")
)
}
\arguments{
\item{data_UniProt_GO}{UniProt data for GO annotations as downloaded by
get_UniProt_data(ids, fields = UniProt_fields()$\verb{Gene Ontology (GO)})}

\item{entry_col}{protein identifier column}

\item{GO_col}{name of column containing GO annotations to be extracted}
}
\description{
Title
}
