% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_interpro.R
\name{get_InterPro_data_from_UniProt}
\alias{get_InterPro_data_from_UniProt}
\title{Title}
\usage{
get_InterPro_data_from_UniProt(
  accession,
  data_UniProt = NULL,
  taxon_ids = NULL,
  max.query = 100,
  keep.empty = F,
  separate.multiple.sites = T,
  split.position = T,
  export_as_UniProt = F,
  silent = T
)
}
\arguments{
\item{accession}{UniProt accession/s to retrieve}

\item{data_UniProt}{(optional) predownloaded UniProt data with
download_UniProt_data() or download_UniProt_data_1o()}

\item{taxon_ids}{taxon identifier/s to restrict the query to}

\item{max.query}{maximum}

\item{keep.empty}{Keep Ids without data?}

\item{separate.multiple.sites}{split disconnected features into multiple rows}

\item{split.position}{add columns 'from' and 'to'}

\item{export_as_UniProt}{return the result shaped like a UniProt table}

\item{silent}{Suppress messages?}
}
\description{
Title
}
\examples{
\dontrun{

}
}
