% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_uniprot.R
\name{get_UniProt_data_o}
\alias{get_UniProt_data_o}
\title{Download UniProt data for given protein accessions, taxonomy identifiers and
data fields (faster for 1000s of proteins; see available fields with
UniProt_fields())}
\usage{
get_UniProt_data_o(
  accession,
  fields = c("accession", "gene_names", "organism_name"),
  taxon_ids = c(human = 9606, mouse = 10090, E.coliK12 = 83333)
)
}
\arguments{
\item{accession}{vector of UniProt accessions}

\item{fields}{UniProt data fields to query}

\item{taxon_ids}{taxonomy Ids}
}
\description{
Download UniProt data for given protein accessions, taxonomy identifiers and
data fields (faster for 1000s of proteins; see available fields with
UniProt_fields())
}
