% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_GO.R
\name{GO_filter_ontology}
\alias{GO_filter_ontology}
\title{Title}
\usage{
GO_filter_ontology(goids, ontology = "CC")
}
\arguments{
\item{goids}{vector of GO identifiers}

\item{ontology}{ontology to keep ("BP", "MF" or "CC")}
}
\description{
Title
}
