% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{max_2}
\alias{max_2}
\title{Calculate maximum along two vectors}
\usage{
max_2(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector 1}

\item{y}{numeric vector 2}

\item{na.rm}{should missing values be removed}
}
\description{
Calculate maximum along two vectors
}
