% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{min_2}
\alias{min_2}
\title{Calculate minimum along two vectors}
\usage{
min_2(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector 1}

\item{y}{numeric vector 2}

\item{na.rm}{should missing values be removed}
}
\description{
Calculate minimum along two vectors
}
