% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{plot_protein_sequence}
\alias{plot_protein_sequence}
\title{Plot PELSA results on a protein sequence including protein domain features}
\usage{
plot_protein_sequence(
  data,
  data_features,
  protein = "P49841",
  protein_range,
  title
)
}
\arguments{
\item{data}{dataset or list holding the peptide data}

\item{data_features}{tibble of sequence features to draw}

\item{protein}{protein group to plot}

\item{protein_range}{residue range to show, as c(start, end)}

\item{title}{plot title}
}
\description{
Plot PELSA results on a protein sequence including protein domain features
}
