% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rev_lvl}
\alias{rev_lvl}
\title{Reverse the factor level order}
\usage{
rev_lvl(x)
}
\arguments{
\item{x}{factor vector}
}
\description{
Reverse the factor level order
}
\examples{
  rev_lvl(factor(c("a", "b", "c")))
}
