% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale_fill_embl}
\alias{scale_fill_embl}
\title{#' ggplot2 'fill' color palette based on EMBL colors from https://www.embl.org/guidelines/design/design-guidelines/colours/}
\usage{
scale_fill_embl(
  palette = c("all", "decorative", "decorative_dark", "decorative_light", "primary",
    "secondary_neutral", "interactive", "ps"),
  sub = c()
)
}
\arguments{
\item{palette}{EMBL color palette}

\item{sub}{vector to choose and order colors}
}
\description{
#' ggplot2 'fill' color palette based on EMBL colors from https://www.embl.org/guidelines/design/design-guidelines/colours/
}
