% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_uniprot.R
\name{UniProt_fields}
\alias{UniProt_fields}
\title{Return list of vectors of UniProt data fields}
\usage{
UniProt_fields()
}
\description{
Return list of vectors of UniProt data fields
}
\examples{
UniProt_fields()
}
