% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_GO.R
\name{UniProt_to_T2GO}
\alias{UniProt_to_T2GO}
\title{Title}
\usage{
UniProt_to_T2GO(
  data_UniProt_GO,
  term_column = "Gene Ontology IDs",
  ontologies = c("CC", "BP", "MF"),
  GO_col = "ids",
  sep = "; "
)
}
\arguments{
\item{data_UniProt_GO}{}

\item{term_column}{}

\item{ontologies}{}

\item{GO_col}{}

\item{sep}{}
}
\description{
Title
}
