% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bin_percent}
\alias{bin_percent}
\title{Title}
\usage{
bin_percent(x, steps = c(0, 0.25, 0.5, 0.75, 1), as_factor = F)
}
\arguments{
\item{x}{vector containing values from 0 to 1}

\item{steps}{}

\item{as_factor}{}
}
\description{
Title
}
\examples{
  bin_percent(seq(0, 1, 0.05))
}
