% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{cheatmap}
\alias{cheatmap}
\title{Title}
\usage{
cheatmap(matrix, draw = T, show_clusters = NULL, add_cluster_n = T, ...)
}
\arguments{
\item{matrix}{}

\item{draw}{}

\item{show_clusters}{}

\item{add_cluster_n}{}

\item{...}{}
}
\description{
Title
}
