% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{.align_axes_limits}
\alias{.align_axes_limits}
\title{Align axis limits of a list of plots}
\usage{
.align_axes_limits(
  list_p,
  x.symmetric = F,
  adjust.x.limits = function(x) x,
  adjust.y.limits = function(y) y,
  ...
)
}
\arguments{
\item{list_p}{list of ggplot objects}

\item{x.symmetric}{set x axis limits to absolute max}

\item{adjust.x.limits}{function or list containing two functions
to be applied to the existing x-axis limits}

\item{adjust.y.limits}{function or list containing two functions
to be applied to the existing y-axis limits}
}
\description{
Align axis limits of a list of plots
}
