% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.cat_contrasts.fit}
\alias{.cat_contrasts.fit}
\title{Creates limma contrasts from design matrix}
\usage{
.cat_contrasts.fit(
  design,
  combinations = tibble(a = c(), b = c()),
  copy2clipboard = T
)
}
\arguments{
\item{design}{design matrix}

\item{combinations}{contrasts to build, as pairs of design columns}

\item{copy2clipboard}{copy the code to the clipboard}
}
\description{
Creates limma contrasts from design matrix
}
\examples{
\dontrun{
  .cat_contrasts.fit(limma_list[["design"]])
}
}
