% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.cat_contrasts.fit}
\alias{.cat_contrasts.fit}
\title{Creates limma contrasts from design matrix}
\usage{
.cat_contrasts.fit(
  design,
  combinations = tibble(a = c(), b = c()),
  copy2clipboard = T
)
}
\arguments{
\item{design}{}

\item{combinations}{}

\item{copy2clipboard}{}
}
\description{
Creates limma contrasts from design matrix
}
\examples{
.cat_contrasts.fit(limma_list[["design"]])
}
