% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{.plot_protein_features}
\alias{.plot_protein_features}
\title{Title}
\usage{
.plot_protein_features(
  data_sites,
  p,
  name = "name",
  start = "from",
  end = "to",
  add.length = 0,
  merge_name_position = F,
  protein_width = 0.6,
  base_size = 16,
  multiple_scaling_factor = 8,
  legend.position = "bottom",
  legend_width = 4,
  colors_manual
)
}
\arguments{
\item{data_sites}{tibble of sequence features to draw}

\item{p}{a ggplot object to add to}

\item{name}{column holding the feature name}

\item{start}{column holding the feature start position}

\item{end}{column holding the feature end position}

\item{add.length}{padding added to each end of the sequence axis}

\item{merge_name_position}{merge features that share a name and position}

\item{protein_width}{height of the protein bar}

\item{base_size}{base font size of the theme}

\item{multiple_scaling_factor}{vertical scaling applied when features overlap}

\item{legend.position}{legend position, passed to ggplot2::theme()}

\item{legend_width}{width of the legend}

\item{colors_manual}{named vector of colours for the features}
}
\description{
Title
}
