% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{.plot_protein_seq}
\alias{.plot_protein_seq}
\title{Title}
\usage{
.plot_protein_seq(
  protein_range = c(1, 100),
  protein_width = 0.6,
  add.length = 0,
  x_lab = "protein",
  y_lab = "",
  color = "#b0b0b0",
  alpha = 0.5,
  base_size = 16
)
}
\arguments{
\item{protein_range}{}

\item{protein_width}{}

\item{add.length}{}

\item{x_lab}{}

\item{y_lab}{}

\item{color}{}

\item{alpha}{}

\item{base_size}{}
}
\description{
Title
}
