% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{.set_PCA_labs}
\alias{.set_PCA_labs}
\title{Title}
\usage{
.set_PCA_labs(p, PCx, PCy, sdev, digits = 1)
}
\arguments{
\item{p}{a ggplot object to add to}

\item{PCx}{principal component shown on the x axis}

\item{PCy}{principal component shown on the y axis}

\item{sdev}{vector of standard deviations from the PCA}

\item{digits}{number of digits shown in the axis labels}
}
\description{
Title
}
