% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelsa_plots.R
\name{plot_protein_sequence}
\alias{plot_protein_sequence}
\title{Plot PELSA results on a protein sequence including protein domain features}
\usage{
plot_protein_sequence(
  data,
  data_features,
  protein = "P49841",
  protein_range,
  title
)
}
\arguments{
\item{data}{}

\item{data_features}{}

\item{protein}{}

\item{protein_range}{}

\item{title}{}
}
\description{
Plot PELSA results on a protein sequence including protein domain features
}
