% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations_GO.R
\name{UniProt_to_T2GO}
\alias{UniProt_to_T2GO}
\title{Title}
\usage{
UniProt_to_T2GO(
  data_UniProt_GO,
  term_column = "Gene Ontology IDs",
  ontologies = c("CC", "BP", "MF"),
  GO_col = "ids",
  sep = "; "
)
}
\arguments{
\item{data_UniProt_GO}{tibble of UniProt records containing a GO column}

\item{term_column}{name of the column holding the term identifiers}

\item{ontologies}{ontologies to keep ("BP", "MF" and/or "CC")}

\item{GO_col}{name of the column holding the GO annotation}

\item{sep}{separator between GO entries within a cell}
}
\description{
Title
}
